Move serde-related stuff to dedicated module
