+++ src/elliptic/curves/wrappers/mod.rs
@@ -7,6 +7,7 @@ mod generator;
 mod point;
 mod point_ref;
 mod scalar;
+mod serde_support;
 
 pub use self::{
     encoded_point::EncodedPoint, encoded_scalar::EncodedScalar, generator::Generator, point::Point,
