+++ src/elliptic/curves/wrappers/scalar.rs
@@ -1,15 +1,8 @@
-use std::marker::PhantomData;
 use std::{fmt, iter};
-
-use serde::de::{Error, MapAccess, Visitor};
-use serde::ser::SerializeStruct;
-use serde::{Deserialize, Serialize};
-use serde::{Deserializer, Serializer};
 
 use crate::elliptic::curves::traits::{Curve, ECScalar};
 use crate::BigInt;
 
-use super::point::CurveNameGuard;
 use crate::elliptic::curves::wrappers::encoded_scalar::EncodedScalar;
 use crate::elliptic::curves::{DeserializationError, ZeroScalarError};
 
@@ -211,148 +204,5 @@ impl<E: Curve> iter::Product for Scalar<E> {
 impl<'s, E: Curve> iter::Product<&'s Scalar<E>> for Scalar<E> {
     fn product<I: Iterator<Item = &'s Scalar<E>>>(iter: I) -> Self {
         iter.fold(Scalar::from(1), |acc, s| acc * s)
-    }
-}
-
-impl<E: Curve> Serialize for Scalar<E> {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let mut s = serializer.serialize_struct("Scalar", 2)?;
-        s.serialize_field("curve", E::CURVE_NAME)?;
-        s.serialize_field(
-            "scalar",
-            // Serializes bytes efficiently
-            serde_bytes::Bytes::new(&self.to_bytes()),
-        )?;
-        s.end()
-    }
-}
-
-impl<'de, E: Curve> Deserialize<'de> for Scalar<E> {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct ScalarVisitor<E: Curve>(PhantomData<E>);
-
-        impl<'de, E: Curve> Visitor<'de> for ScalarVisitor<E> {
-            type Value = Scalar<E>;
-
-            fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
-                write!(f, "scalar of {} curve", E::CURVE_NAME)
-            }
-
-            fn visit_map<A>(self, mut map: A) -> Result<Self::Value, A::Error>
-            where
-                A: MapAccess<'de>,
-            {
-                let mut curve_name: Option<CurveNameGuard<E>> = None;
-                let mut scalar: Option<ScalarFromBytes<E>> = None;
-
-                while let Some(key) = map.next_key()? {
-                    match key {
-                        Field::Curve => {
-                            if curve_name.is_some() {
-                                return Err(A::Error::duplicate_field("curve_name"));
-                            }
-                            curve_name = Some(map.next_value()?)
-                        }
-                        Field::Scalar => {
-                            if scalar.is_some() {
-                                return Err(A::Error::duplicate_field("scalar"));
-                            }
-                            scalar = Some(map.next_value()?)
-                        }
-                    }
-                }
-                let _curve_name =
-                    curve_name.ok_or_else(|| A::Error::missing_field("curve_name"))?;
-                let scalar = scalar.ok_or_else(|| A::Error::missing_field("scalar"))?;
-                Ok(scalar.0)
-            }
-        }
-
-        deserializer.deserialize_struct("Scalar", &["curve", "scalar"], ScalarVisitor(PhantomData))
-    }
-}
-
-struct ScalarFromBytes<E: Curve>(Scalar<E>);
-
-impl<'de, E: Curve> Deserialize<'de> for ScalarFromBytes<E> {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct ScalarBytesVisitor<E: Curve>(PhantomData<E>);
-
-        impl<'de, E: Curve> Visitor<'de> for ScalarBytesVisitor<E> {
-            type Value = Scalar<E>;
-
-            fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
-                write!(f, "scalar value of {} curve", E::CURVE_NAME)
-            }
-
-            fn visit_bytes<Err>(self, v: &[u8]) -> Result<Self::Value, Err>
-            where
-                Err: Error,
-            {
-                Scalar::from_bytes(v).map_err(|_| Err::custom("invalid scalar"))
-            }
-        }
-
-        deserializer
-            .deserialize_bytes(ScalarBytesVisitor(PhantomData))
-            .map(ScalarFromBytes)
-    }
-}
-
-#[derive(Deserialize)]
-#[serde(field_identifier, rename_all = "snake_case")]
-enum Field {
-    Curve,
-    Scalar,
-}
-
-#[cfg(test)]
-mod serde_tests {
-    use serde_test::{assert_tokens, Token::*};
-
-    use crate::elliptic::curves::*;
-
-    #[test]
-    fn test_serde_scalar() {
-        fn generic<E: Curve>(scalar: Scalar<E>) {
-            let bytes = scalar.to_bytes().to_vec();
-            let tokens = vec![
-                Struct {
-                    name: "Scalar",
-                    len: 2,
-                },
-                Str("curve"),
-                Str(E::CURVE_NAME),
-                Str("scalar"),
-                Bytes(bytes.leak()),
-                StructEnd,
-            ];
-            assert_tokens(&scalar, &tokens);
-        }
-
-        // Test **zero scalars** (de)serializing
-        generic::<Secp256k1>(Scalar::zero());
-        generic::<Secp256r1>(Scalar::zero());
-        generic::<Ed25519>(Scalar::zero());
-        generic::<Ristretto>(Scalar::zero());
-        generic::<Bls12_381_1>(Scalar::zero());
-        generic::<Bls12_381_2>(Scalar::zero());
-
-        // Test **random scalars** (de)serializing
-        generic::<Secp256k1>(Scalar::random());
-        generic::<Secp256r1>(Scalar::random());
-        generic::<Ed25519>(Scalar::random());
-        generic::<Ristretto>(Scalar::random());
-        generic::<Bls12_381_1>(Scalar::random());
-        generic::<Bls12_381_2>(Scalar::random());
     }
 }
